Adjusted the Cache ResultModifier to the updated codebase
